crud_service/requirements.txt
flask==1.1.2
mongoengine==0.20.0
flask_restful
flask_mongoengine
dnspython
gunicorn
requests
python-dateutil
pandas==1.0.4
kafka-python